requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Pillow==9.0.0
Flask-OAuthlib==0.9.6
